refactor(lib): add `lib.rs` and export `run` function
